shoe_wear_project_traincode/requirements.txt
# ---------------------------
# PyTorch + Vision
# ---------------------------
torch==2.2.0
torchvision==0.17.0

# ---------------------------
# 필수 유틸
# ---------------------------
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
tqdm==4.66.2
Pillow==10.2.0

# ---------------------------
# 데이터 처리 / 이미지 관련
# ---------------------------
opencv-python==4.9.0.80

# ---------------------------
# 모델 학습 관련
# ---------------------------
matplotlib==3.8.2

# ---------------------------
# 기타 필요 라이브러리
# ---------------------------
pyyaml==6.0.1
